seperated users and user endpoints + search for user only data
