+++ backend/src/user/controllers.ts
@@ -1,0 +1,48 @@
+import { Prisma } from "@prisma/client";
+import { Request, Response } from "express";
+
+// We are using one connection to prisma client to prevent multiple connections
+import prisma from "../../client";
+
+/**
+ * Finds all users in DB
+ * @returns promise with all users or error
+ */
+const getUsers = async (req: Request, res: Response) => {
+  // #swagger.tags = ['Users']
+  try {
+    const users = await prisma.user.findMany();
+    res.status(200).json(users);
+  } catch (error: any) {
+    // Do we need to log the error?
+    res.status(500).json({ error: error.message });
+  }
+};
+
+/**
+ * Sign up a new user. User details are passed in request body.
+ * Request body includes:
+ * - email (string)
+ * - name (string)
+ * @returns promise with user or error
+ */
+const createNewUser = async (req: Request, res: Response) => {
+  // #swagger.tags = ['Users']
+  try {
+    const result = await prisma.user.create({
+      data: {
+        ...req.body,
+      },
+    });
+
+    res.status(201).json(result);
+  } catch (err: any) {
+    res.status(500).json({ error: err.message });
+  }
+};
+
+
+export default {
+  getUsers,
+  createNewUser,
+};
