+++ backend/src/users/views.ts
@@ -1,14 +1,12 @@
 import { Router } from "express";
-import userController from "./controllers";
+import usersController from "./controllers";
 
-const userRouter = Router();
+const usersRouter = Router();
 
 /** User Specific Routes */
 
 // This approach is cleaner for us becuase we can easily add middle ware
-userRouter.get("/", userController.getUsers);
-userRouter.post("/signup", userController.createNewUser);
-userRouter.get("/all", userController.getAllUsers);
-userRouter.get("/search", userController.getSearchedUser);
+usersRouter.get("/all", usersController.getAllUsers);
+usersRouter.get("/search", usersController.getSearchedUser);
 
-export default userRouter;
+export default usersRouter;
